feat(build): sync quasar-ui version to quasar-app-extension
The new code will look for both app-extension and ui and on a build. If both are found, it will sync the ui version in dependencies/devDependencies. Optionally (based n setting), will also sync the app-extension version to be the same as the ui version. This is based on tedious take of manually syncing, of which has been missed in the past by several developers, to streamline this process.